requirements.txt
djangorestframework
drf-yasg
psycopg2-binary
djangorestframework-simplejwt
Pillow